Do not generate `GetBy[Unique]()` columns if the column is nullable
Columns that are both unique and nullable are allowed, but because
SQLite treats all NULLs as unique, querying for NULL can return multiple
rows, as opposed to one which we expect from a `GetBy[Unique]()` getter.
To avoid this confusion and potential issue, I updated the logic to
simply not generate these functions.

If this is unsatisfactory, we could also generate two getters, one for
querying for NULL and one for querying by a non-null value.