BUG: Ignore `CREATE [UNIQUE] INDEX ...` including if over multiple lines
The first bug was simply that the previous parser did not support the
optional `UNIQUE` keyword when parsing indices. This was partially
addressed in the previous commit by parsing it, but even this wasn't
great because it didn't cover all possible index expressions or indices
on multiple lines. I "solved" this by finding the closing semi-colon and
ignoring the entire statement. I also added logic to handle new lines
for indicies.

Two unit tests were added to cover these two bugs.